Stop using system include files
There's no need to use system include headers especially considering that we're not even linking to the libc that they describe.
Instead, roll our own `common.h`, and use the `-include` flag to automatically include it in every file.

+++ include/gpio.h
@@ -1,10 +1,6 @@
 #pragma once
 
-#include <stdbool.h>
-#include <stddef.h>
-#include <stdint.h>
-
-typedef unsigned int gpio_pin_t;
+typedef u8 gpio_pin_t;
 
 typedef enum gpio_mode {
 	gpio_mode_input = 0,
